A moderatly well working (1 server) example
The Operation class has changed ; it no longer holds a OperationType
enum, because it needed to be Serializable. Although it would have
been 'easy' to make the enum and the class Serializable, since we
only have two operations, it is way easier to just make it an int
and to check if the value is 0 or 1.

An operation of type 0 is FIB and 1 is PRIME.

Also has a bit more info on the outputs...